fix(position): fix position calculation for dropdown and extract util function
