+++ src/dialog/overflow-menu/overflow-menu-pane.component.ts
@@ -12,6 +12,7 @@ import { I18n } from "carbon-components-angular/i18n";
 import { ExperimentalService } from "carbon-components-angular/experimental";
 import { ElementService } from "carbon-components-angular/utils";
 import { CloseReasons } from "../dialog-config.interface";
+import { closestAttr } from '../../utils/position';
 
 /**
  * Extend the `Dialog` component to create an overflow menu.
@@ -55,7 +56,7 @@ export class OverflowMenuPane extends Dialog implements AfterViewInit {
 			* we also move the element by half of it's own width, since
 			* position service will try and center everything
 			*/
-			const closestRel = this.closestAttr("position", ["relative", "fixed", "absolute"]);
+			const closestRel = closestAttr("position", ["relative", "fixed", "absolute"], this.elementRef.nativeElement);
 			const topFix = closestRel ? closestRel.getBoundingClientRect().top * -1 : 0;
 			const leftFix = closestRel ? closestRel.getBoundingClientRect().left * -1 : 0;
 
@@ -73,25 +74,6 @@ export class OverflowMenuPane extends Dialog implements AfterViewInit {
 		if (!this.dialogConfig.menuLabel) {
 			this.dialogConfig.menuLabel = this.i18n.get().OVERFLOW_MENU.OVERFLOW;
 		}
-	}
-
-	closestAttr(s, t) {
-		let el = this.elementRef.nativeElement;
-
-		do {
-			if (
-				this.matchesAttr(el, s, t)
-			) {
-				return el;
-			}
-			el = el.parentElement || el.parentNode;
-		} while (el !== null && el.nodeType === 1);
-		return null;
-	}
-
-	matchesAttr(el, attr, val) {
-		const styles = window.getComputedStyle(el);
-		return val.includes(styles[attr]);
 	}
 
 	@HostListener("keydown", ["$event"])
